+++ rocker-runtime/src/main/java/com/fizzed/rocker/runtime/Elvis.java
@@ -1,0 +1,9 @@
+package com.fizzed.rocker.runtime;
+
+public class Elvis {
+    
+    static public Object op(Object a, Object b) {
+        return (a != null ? a : b);
+    }
+    
+}